Split module into adapter and algorithm
