+++ algorithms/sunTrackError/sunTrackError.h
@@ -1,6 +1,7 @@
 #ifndef F32XMERA_SUN_TRACK_ERROR_H
 #define F32XMERA_SUN_TRACK_ERROR_H
 
+#include "sunTrackErrorAlgorithm.h"
 #include <architecture/messaging/messaging.h>
 #include <architecture/msgPayloadDef/AttGuidMsgPayload.h>
 #include <architecture/msgPayloadDef/AttRefMsgPayload.h>
@@ -12,14 +13,14 @@
 
 /*!@brief Module to compute the attitude tracking error for sun avoidance.
  */
-class SunTrackError : public SysModel {
+class SunTrackError final : public SysModel {
    public:
+    SunTrackError() = default;
+    ~SunTrackError() override = default;
+
     void reset(uint64_t callTime) override;
     void updateState(uint64_t callTime) override;
 
-    AttGuidMsgPayload computeSunTrackError(NavAttMsgPayload& nav,
-                                           AttRefMsgPayload& ref,
-                                           uint64_t callTime) const;
     void setSigma_R0R(const Eigen::Vector3d& sigma);
     Eigen::Vector3d getSigma_R0R() const;
     void setSensitiveHat_B(const Eigen::Vector3d& sensitiveDirection);
@@ -34,15 +35,7 @@ class SunTrackError : public SysModel {
     Message<AttGuidMsgPayload> attGuidOutMsg;         //!< output msg of attitude guidance
 
    private:
-    Eigen::Vector3d sigma_R0R{Eigen::Vector3d::Zero()};  /*!< MRP from corrected reference frame to original frame R0
-                                                            This is the same as [BcB] going from primary body frame B
-                                                            to the corrected body frame Bc */
-    Eigen::Vector3d sensitiveHat_B{Eigen::Vector3d::Zero()};  //!< [-] Vehicle body vector to exclude from sun*/
-    double angleRate{};                //!< [r/s] The rate at which we maneuver to Sun point*/
-    Eigen::Vector3d mnvrAxis_B{Eigen::Vector3d::Zero()};  //!< [-] Eigen axis that we are maneuvering on*/
-    double angleStart{};           //!< [r] The angle remaining in the attitude maneuver*/
-    bool maneuverInitialized{};     //!< [-] Flag indicating if maneuver has been set*/
-    uint64_t mnvrStartTime{};      //!< [ns] Time at which the maneuver was begun*/
+    SunTrackErrorAlgorithm algorithm{};
 };
 
 #endif
